+++ packages/interactivity-router/src/assets/script-modules.ts
@@ -2,29 +2,11 @@
  * Internal dependencies
  */
 import {
+	initialImportMap,
 	importPreloadedModule,
 	preloadWithMap,
 	type ModuleLoad,
 } from './dynamic-importmap';
-
-/**
- * Script element containing the initial page's import map.
- */
-const initialImportMapElement =
-	window.document.querySelector< HTMLScriptElement >(
-		'script#wp-importmap[type=importmap]'
-	);
-
-/**
- * Data from the initial page's import map.
- *
- * Pages containing any of the imports present on the original page
- * in their import maps should ignore them, as those imports would
- * be handled natively.
- */
-const initialImportMap = initialImportMapElement
-	? JSON.parse( initialImportMapElement.text )
-	: { imports: {}, scopes: {} };
 
 /**
  * IDs of modules that should be resolved by the browser rather than